Delete TimeWatchedWinRatioAndProfitRatioLoss.py
It might not make sense to optimize for shorter trade durations at hyper-opt level: if it is indeed more efficient to stop a trade and keep it short, it would eventually manifests itself in the 'OnlyProfitLoss' loss function from FreqTrade, or in the MGM-custom 'WinRatioAndAbsProfitLoss' one to be kept in this feature branch. 

Another way to say this: keeping trade short will be a BY-PRODUCT of hyperopting for best profits, since the shorter a trade is kept, even if winning, the more trades can be done and potentially result in more profits, which is what 'OnlyProfitLoss' and 'WinRatioAndAbsProfitLoss' already optimize for.